requirements.txt
pyramid==1.8.1
cornice==2.4.0
SQLAlchemy==1.1.5
psycopg2==2.6.2
pyjwt==1.4.2
pytest==3.0.6
waitress==1.0.2